Fix #815: remove dependency on cats-effect from monix-execution
