+++ monix-catnap/jvm/src/test/scala_2.12/monix/catnap/FutureLiftJava8Suite.scala
@@ -44,7 +44,7 @@ object FutureLiftJava8Suite extends TestSuite[TestScheduler] {
   }
 
   test("convert from async CompletableFuture; on success; with Concurrent[IO]") { implicit s =>
-    implicit val cs = s.contextShift[IO]
+    implicit val cs = SchedulerEffect.contextShift[IO](s)
 
     val future = new CompletableFuture[Int]()
     val f = IO(future).futureLift.unsafeToFuture()
@@ -72,7 +72,7 @@ object FutureLiftJava8Suite extends TestSuite[TestScheduler] {
   }
 
   test("convert from async CompletableFuture; on failure; with Concurrent[IO]") { implicit s =>
-    implicit val cs = s.contextShift[IO]
+    implicit val cs = SchedulerEffect.contextShift[IO](s)
 
     val future = new CompletableFuture[Int]()
     val f = convertConcurrent(IO(future)).unsafeToFuture()
@@ -88,7 +88,7 @@ object FutureLiftJava8Suite extends TestSuite[TestScheduler] {
   }
 
   test("CompletableFuture is cancelable via IO") { implicit s =>
-    implicit val cs = s.contextShift[IO]
+    implicit val cs = SchedulerEffect.contextShift[IO](s)
 
     val future = new CompletableFuture[Int]()
