+++ monix-catnap/shared/src/main/scala/monix/catnap/ConcurrentQueue.scala
@@ -44,9 +44,9 @@ import scala.collection.mutable.ArrayBuffer
   *   import monix.execution.Scheduler.global
   *
   *   // For being able to do IO.start
-  *   implicit val cs = global.contextShift[IO]
+  *   implicit val cs = SchedulerEffect.contextShift[IO](global)
   *   // We need a `Timer` for this to work
-  *   implicit val timer = global.timer[IO]
+  *   implicit val timer = SchedulerEffect.timer[IO](global)
   *
   *   def consumer(queue: ConcurrentQueue[IO, Int], index: Int): IO[Unit] =
   *     queue.poll.flatMap { a =>
