+++ monix-catnap/shared/src/test/scala/monix/execution/CallbackInstanceSuite.scala
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2014-2019 by The Monix Project Developers.
+ * See the project homepage at: https://monix.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package monix.execution
+import cats.Contravariant
+import minitest.TestSuite
+import monix.execution.CallbackSuite.TestCallback
+import monix.execution.schedulers.TestScheduler
+
+object CallbackInstanceSuite extends TestSuite[TestScheduler] {
+  def setup() = TestScheduler()
+  def tearDown(env: TestScheduler): Unit =
+    assert(env.state.tasks.isEmpty, "should not have tasks left to execute")
+
+  test("contramap has a cats Contramap instance") { implicit s =>
+    val instance = implicitly[Contravariant[Callback[Throwable, ?]]]
+    val callback = TestCallback()
+    val stringCallback = instance.contramap(callback)((x: String) => x.toInt)
+    stringCallback.onSuccess("1")
+    assert(callback.successCalled)
+  }
+}
