+++ monix-eval/shared/src/main/scala/monix/eval/TaskApp.scala
@@ -18,6 +18,7 @@
 package monix.eval
 
 import cats.effect._
+import monix.catnap.SchedulerEffect
 import monix.eval.instances.CatsConcurrentEffectForTask
 import monix.execution.Scheduler
 
@@ -86,9 +87,9 @@ trait TaskApp {
     val self = this
     val app = new IOApp {
       override implicit lazy val contextShift: ContextShift[IO] =
-        scheduler.contextShift[IO](IO.ioEffect)
+        SchedulerEffect.contextShift[IO](scheduler)(IO.ioEffect)
       override implicit lazy val timer: Timer[IO] =
-        scheduler.timerLiftIO[IO](IO.ioEffect)
+        SchedulerEffect.timerLiftIO[IO](scheduler)(IO.ioEffect)
       def run(args: List[String]): IO[ExitCode] =
         self.run(args).toIO
     }
