+++ monix-execution/shared/src/main/scala/monix/execution/Cancelable.scala
@@ -17,9 +17,8 @@
 
 package monix.execution
 
-import cats.effect.{CancelToken, IO, Sync}
 import monix.execution.atomic.AtomicAny
-import monix.execution.internal.{AttemptCallback, Platform}
+import monix.execution.internal. Platform
 
 import scala.util.control.NonFatal
 import monix.execution.schedulers.TrampolinedRunnable
@@ -97,31 +96,6 @@ object Cancelable {
         p.tryFailure(e)
     }
 
-  /** Builds a [[Cancelable]] reference from an `IO[Unit]`.
-    *
-    * Guarantees idempotency and reports any uncaught errors.
-    *
-    * @param io is the `IO` value to evaluate on `cancel`
-    * @param r is an exception reporter that's used in case our `IO`
-    *        value is throwing an error on evaluation
-    */
-  def fromIO(io: IO[Unit])(implicit r: UncaughtExceptionReporter): Cancelable =
-    Cancelable { () =>
-      io.unsafeRunAsync(AttemptCallback.empty)
-    }
-
-  /** Internal API — builds a `Cancelable` reference from an `IO[Unit]`,
-    * but without any protections for idempotency.
-    */
-  private[monix] def fromIOUnsafe(io: IO[Unit])
-    (implicit r: UncaughtExceptionReporter): Cancelable = {
-
-    new Cancelable {
-      def cancel(): Unit =
-        io.unsafeRunAsync(AttemptCallback.empty)
-    }
-  }
-
   /** Given a collection of cancelables, cancel them all.
     *
     * This function collects non-fatal exceptions and throws them all
@@ -153,20 +127,6 @@ object Cancelable {
 
   /** Marker for cancelables that are dummies that can be ignored. */
   trait IsDummy { self: Cancelable => }
-
-  /** Extension methods for [[Cancelable]]. */
-  implicit final class Extensions(val self: Cancelable) extends AnyVal {
-    /**
-      * Given a [[Cancelable]] reference, turn it into an
-      * `CancelToken[F]` (a Cats-Effect alias for `F[Unit]`) that
-      * will trigger [[Cancelable.cancel cancel]] on evaluation.
-      */
-    def toCancelToken[F[_]](implicit F: Sync[F]): CancelToken[F] =
-      self match {
-        case _: IsDummy => F.unit
-        case _ => F.delay(self.cancel())
-      }
-  }
 
   private final class CancelableTask(cb: () => Unit)
     extends Cancelable {
