+++ monix-execution/shared/src/main/scala/monix/execution/Scheduler.scala
@@ -18,9 +18,6 @@
 package monix.execution
 
 import java.util.concurrent.Executor
-import cats.implicits._
-import cats.effect._
-import monix.execution.internal.AttemptCallback.RunnableTick
 import monix.execution.internal.RunnableAction
 import monix.execution.schedulers.SchedulerCompanionImpl
 import scala.annotation.implicitNotFound
@@ -254,97 +251,6 @@ object Scheduler extends SchedulerCompanionImpl {
 
   /** Utilities complementing the `Scheduler` interface. */
   implicit final class Extensions(val source: Scheduler) extends AnyVal with schedulers.ExecuteExtensions {
-
-    /**
-      * Derives a `cats.effect.Clock` from [[Scheduler]] for any
-      * data type that has a `cats.effect.LiftIO` implementation.
-      */
-    def clock[F[_]](implicit F: Sync[F]): Clock[F] =
-      new Clock[F] {
-        override def realTime(unit: TimeUnit): F[Long] =
-          F.delay(source.clockRealTime(unit))
-        override def monotonic(unit: TimeUnit): F[Long] =
-          F.delay(source.clockMonotonic(unit))
-      }
-
-    /**
-      * Derives a `cats.effect.Timer` from [[Scheduler]] for any
-      * data type that has a `cats.effect.Concurrent` type class
-      * instance.
-      *
-      * {{{
-      *   implicit val timer: Timer[IO] = scheduler.timer[IO]
-      *
-      *   IO.sleep(10.seconds).flatMap { _ =>
-      *     IO(println("Delayed hello!"))
-      *   }
-      * }}}
-      */
-    def timer[F[_]](implicit F: Concurrent[F]): Timer[F] =
-      new Timer[F] {
-        override def sleep(d: FiniteDuration): F[Unit] =
-          F.cancelable { cb =>
-            source.scheduleOnce(d.length, d.unit, new RunnableTick(cb))
-              .toCancelToken[F]
-          }
-        override def clock: Clock[F] =
-          source.clock
-      }
-
-    /**
-      * Derives a `cats.effect.Timer` from [[Scheduler]] for any
-      * data type that has a `cats.effect.LiftIO` instance.
-      *
-      * This is the relaxed [[timer]] method, needing only `LiftIO`
-      * to work, by piggybacking on `cats.effect.IO`.
-      *
-      * {{{
-      *   implicit val timer: Timer[IO] = scheduler.timerLiftIO[IO]
-      *
-      *   IO.sleep(10.seconds).flatMap { _ =>
-      *     IO(println("Delayed hello!"))
-      *   }
-      * }}}
-      */
-    def timerLiftIO[F[_]](implicit F: LiftIO[F]): Timer[F] =
-      new Timer[F] {
-        override def sleep(d: FiniteDuration): F[Unit] =
-          F.liftIO(IO.cancelable { cb =>
-            source.scheduleOnce(d.length, d.unit, new RunnableTick(cb))
-              .toCancelToken[IO]
-          })
-        override def clock: Clock[F] =
-          new Clock[F] {
-            def realTime(unit: TimeUnit): F[Long] =
-              F.liftIO(IO(source.clockRealTime(unit)))
-            def monotonic(unit: TimeUnit): F[Long] =
-              F.liftIO(IO(source.clockMonotonic(unit)))
-          }
-      }
-
-    /**
-      * Derives a `cats.effect.ContextShift` from [[Scheduler]] for any
-      * data type that has a `cats.effect.Effect` implementation.
-      *
-      * {{{
-      *   val contextShift: ContextShift[IO] = scheduler.contextShift[IO]
-      *   val executor = Executors.newCachedThreadPool()
-      *   val ec = ExecutionContext.fromExecutor(executor)
-      *
-      *   contextShift.evalOn(ec)(IO(println("I'm on different thread pool!"))
-      *     .flatMap { _ =>
-      *       IO(println("I came back to default"))
-      *     }
-      * }}}
-      */
-    def contextShift[F[_]](implicit F: Async[F]): ContextShift[F] =
-      new ContextShift[F] {
-        override def shift: F[Unit] =
-          Async.shift(source)
-        override def evalOn[A](ec: ExecutionContext)(fa: F[A]): F[A] =
-          Async.shift(ec).flatMap(_ => fa.flatMap(a => shift.map(_ => a)))
-      }
-
     /** Schedules a task to run in the future, after `initialDelay`.
       *
       * For example the following schedules a message to be printed to
