+++ monix-execution/shared/src/test/scala/monix/execution/CancelableSuite.scala
@@ -137,50 +137,4 @@ object CancelableSuite extends SimpleTestSuite {
     c.cancel()
     assertEquals(p.future.value, Some(Failure(dummy)))
   }
-
-  test("Cancelable.fromIO") {
-    implicit val ctx = TestScheduler()
-
-    var effect = 0
-    val io = IO {
-      effect += 1
-    }
-    val c = Cancelable.fromIO(io)
-
-    assertEquals(effect, 0)
-    c.cancel()
-    assertEquals(effect, 1)
-    c.cancel()
-    assertEquals(effect, 1)
-  }
-
-  test("Cancelable.fromIO reports error") {
-    implicit val ctx = TestScheduler()
-    val dummy = DummyException("dummy")
-
-    val io = IO {
-      throw dummy
-    }
-    val c = Cancelable.fromIO(io)
-
-    assertEquals(ctx.state.lastReportedError, null)
-    c.cancel()
-    assertEquals(ctx.state.lastReportedError, dummy)
-  }
-
-  test("Cancelable#toCancelToken") {
-    var effect = 0
-    val c = Cancelable { () => effect += 1 }
-    val io = c.toCancelToken[IO]
-
-    assertEquals(effect, 0)
-    io.unsafeRunSync()
-    assertEquals(effect, 1)
-    io.unsafeRunSync()
-    assertEquals(effect, 1)
-  }
-
-  test("Cancelable.empty.toCancelToken[IO] == IO.unit") {
-    assertEquals(Cancelable.empty.toCancelToken[IO], IO.unit)
-  }
 }
