+++ monix-execution/shared/src/test/scala/monix/execution/schedulers/ReferenceSchedulerSuite.scala
@@ -144,62 +144,4 @@ object ReferenceSchedulerSuite extends SimpleTestSuite {
     s.tick()
     assertEquals(s.underlying.state.lastReportedError, dummy)
   }
-
-  test("clock.monotonic") {
-    val s = new DummyScheduler
-    val clock = s.clock[IO]
-
-    val clockMonotonic = clock.monotonic(MILLISECONDS).unsafeRunSync()
-    assert(clockMonotonic > 0)
-  }
-
-  test("clock.realTime") {
-    val s = new DummyScheduler
-    val clock = s.clock[IO]
-
-    val clockRealTime = clock.realTime(MILLISECONDS).unsafeRunSync()
-    assert(clockRealTime > 0)
-  }
-
-  test("timer.sleep") {
-    val s = new DummyScheduler
-    val timer = s.timerLiftIO[IO]
-
-    val f = timer.sleep(10.seconds).unsafeToFuture()
-    assertEquals(f.value, None)
-
-    s.tick(5.seconds)
-    assertEquals(f.value, None)
-
-    s.tick(5.seconds)
-    assertEquals(f.value, Some(Success(())))
-  }
-
-  test("contextShift.shift") {
-    val s = new DummyScheduler
-    val contextShift = s.contextShift[IO]
-
-    val f = contextShift.shift.unsafeToFuture()
-    assertEquals(f.value, None)
-
-    s.tick()
-    assertEquals(f.value, Some(Success(())))
-  }
-
-  test("contextShift.evalOn") {
-    val s = new DummyScheduler
-    val contextShift = s.contextShift[IO]
-    val s2 = new DummyScheduler()
-
-    val f = contextShift.evalOn(s2)(IO(1)).unsafeToFuture()
-    assertEquals(f.value, None)
-
-    s.tick()
-    assertEquals(f.value, None)
-
-    s2.tick()
-    assertEquals(f.value, None)
-    s.tick()
-    assertEquals(f.value, Some(Success(1)))
-  }
 }
