+++ monix-tail/shared/src/test/scala/monix/tail/IntervalIntervalSuite.scala
@@ -22,6 +22,7 @@ import monix.eval.Task
 
 import scala.util.Success
 import scala.concurrent.duration._
+import monix.catnap.SchedulerEffect
 
 object IntervalIntervalSuite extends BaseTestSuite {
   test("Iterant[Task].intervalWithFixedDelay(1.second, 2.seconds)") { implicit s =>
@@ -54,7 +55,7 @@ object IntervalIntervalSuite extends BaseTestSuite {
   }
 
   test("Iterant[IO].intervalWithFixedDelay(1.second, 2.seconds)") { s =>
-    implicit val timer = s.timerLiftIO[IO]
+    implicit val timer = SchedulerEffect.timerLiftIO[IO](s)
 
     var effect = 0
     val lst = Iterant[IO].intervalWithFixedDelay(1.second, 2.seconds)
@@ -108,7 +109,7 @@ object IntervalIntervalSuite extends BaseTestSuite {
   }
 
   test("Iterant[IO].intervalWithFixedDelay(2.seconds)") { s =>
-    implicit val timer = s.timerLiftIO[IO]
+    implicit val timer = SchedulerEffect.timerLiftIO[IO](s)
 
     var effect = 0
     val lst = Iterant[IO].intervalWithFixedDelay(2.seconds)
@@ -155,7 +156,7 @@ object IntervalIntervalSuite extends BaseTestSuite {
   }
 
   test("Iterant[IO].intervalAtFixedRate(1.second)") { s =>
-    implicit val timer = s.timerLiftIO[IO]
+    implicit val timer = SchedulerEffect.timerLiftIO[IO](s)
 
     var effect = 0
     val lst = Iterant[IO].intervalAtFixedRate(1.second)
@@ -203,7 +204,7 @@ object IntervalIntervalSuite extends BaseTestSuite {
   }
 
   test("Iterant[IO].intervalAtFixedRate(2.seconds, 1.second)") { s =>
-    implicit val timer = s.timerLiftIO[IO]
+    implicit val timer = SchedulerEffect.timerLiftIO[IO](s)
 
     var effect = 0
     val lst = Iterant[IO].intervalAtFixedRate(2.seconds, 1.second)
@@ -252,7 +253,7 @@ object IntervalIntervalSuite extends BaseTestSuite {
   }
 
   test("Iterant[IO].intervalAtFixedRate accounts for time it takes task to finish") { s =>
-    implicit val timer = s.timerLiftIO[IO]
+    implicit val timer = SchedulerEffect.timerLiftIO[IO](s)
 
     var effect = 0
     val lst = Iterant[IO].intervalAtFixedRate(1.second)
