+++ monix-tail/shared/src/test/scala/monix/tail/IterantChannelSuite.scala
@@ -25,14 +25,15 @@ import monix.execution.BufferCapacity.{Bounded, Unbounded}
 import monix.execution.ChannelType.{MultiProducer, SingleProducer}
 import monix.execution.internal.Platform
 import monix.execution.{BufferCapacity, Scheduler}
+import monix.catnap.SchedulerEffect
 
 object IterantChannelSuite extends SimpleTestSuite {
   implicit val ec: Scheduler = Scheduler.global
 
   implicit def contextShift(implicit s: Scheduler): ContextShift[IO] =
-    s.contextShift[IO](IO.ioEffect)
+    SchedulerEffect.contextShift[IO](s)(IO.ioEffect)
   implicit def timer(implicit s: Scheduler): Timer[IO] =
-    s.timerLiftIO[IO](IO.ioEffect)
+    SchedulerEffect.timerLiftIO[IO](s)(IO.ioEffect)
 
   def testIO(name: String)(f: => IO[Unit]) =
     testAsync(name)(f.unsafeToFuture())
